Move NonRandom into pipeline_utils and setup v2
